updated microphone permissions in dashboard fragment
