+++ org.envirocar.app/src/org/envirocar/app/views/dashboard/DashboardFragment.java
@@ -19,6 +19,8 @@
 package org.envirocar.app.views.dashboard;
 
 import static android.app.Activity.RESULT_OK;
+import static org.envirocar.app.views.utils.SnackbarUtil.showGrantMicrophonePermission;
+import static org.envirocar.app.views.utils.SnackbarUtil.showSnackbarLong;
 import static org.envirocar.app.views.utils.SnackbarUtil.showVoiceTriggeredSnackbar;
 
 import android.Manifest;
@@ -59,7 +61,6 @@ import androidx.appcompat.widget.Toolbar;
 import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.constraintlayout.widget.ConstraintSet;
 import androidx.core.app.ActivityCompat;
-import androidx.lifecycle.ViewModelProvider;
 
 import com.afollestad.materialdialogs.MaterialDialog;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
@@ -75,7 +76,6 @@ import com.google.android.play.core.tasks.Task;
 import com.jakewharton.rxbinding3.appcompat.RxToolbar;
 import com.justai.aimybox.Aimybox;
 import com.justai.aimybox.components.AimyboxAssistantViewModel;
-import com.justai.aimybox.components.AimyboxProvider;
 import com.squareup.otto.Subscribe;
 
 import org.envirocar.app.BaseApplicationComponent;
@@ -91,7 +91,6 @@ import org.envirocar.app.recording.RecordingState;
 import org.envirocar.app.recording.RecordingType;
 import org.envirocar.app.recording.events.EngineNotRunningEvent;
 import org.envirocar.app.recording.events.RecordingStateEvent;
-import org.envirocar.app.views.BaseMainActivity;
 import org.envirocar.app.views.carselection.CarSelectionActivity;
 import org.envirocar.app.views.login.SigninActivity;
 import org.envirocar.app.views.obdselection.OBDSelectionActivity;
@@ -111,12 +110,12 @@ import org.envirocar.core.utils.ServiceUtils;
 import org.envirocar.obd.events.TrackRecordingServiceStateChangedEvent;
 import org.envirocar.obd.service.BluetoothServiceState;
 import org.envirocar.voicecommand.BaseAimybox;
+import org.envirocar.voicecommand.BaseAimyboxAssistantViewModel;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Objects;
 import java.util.concurrent.TimeUnit;
 
 import javax.inject.Inject;
@@ -135,14 +134,17 @@ import kotlin.coroutines.CoroutineContext;
 import kotlinx.coroutines.CoroutineScope;
 import kotlinx.coroutines.Dispatchers;
 import kotlinx.coroutines.JobKt;
+import pub.devrel.easypermissions.EasyPermissions;
+import pub.devrel.easypermissions.PermissionRequest;
 
 /**
  * @author dewall
  */
-public class DashboardFragment extends BaseInjectorFragment implements CoroutineScope {
+public class DashboardFragment extends BaseInjectorFragment implements CoroutineScope, EasyPermissions.PermissionCallbacks {
     private static final Logger LOG = Logger.getLogger(DashboardFragment.class);
 
     private static final int LOCATION_PERMISSION_REQUEST_CODE = 1203;
+    public static final int RECORD_AUDIO_PERMISSION_REQ_CODE = 55;
 
     // View Injections
     @BindView(R.id.fragment_dashboard_toolbar)
@@ -290,37 +292,29 @@ public class DashboardFragment extends BaseInjectorFragment implements Coroutine
     @Override
     public void onAttach(Context context) {
         super.onAttach(context);
-
-        AimyboxProvider aimyboxProvider = Objects.requireNonNull(
-                new BaseAimybox().findAimyboxProvider(requireActivity()),
-                "Parent Activity or Application must implement AimyboxProvider interface"
-        );
-
-        if (viewModel == null) {
-            viewModel =
-                    new ViewModelProvider(requireActivity(), aimyboxProvider.getViewModelFactory())
-                            .get(AimyboxAssistantViewModel.class);
-
-            new BaseAimybox().Companion.setInitialPhrase(context, getArguments(), viewModel);
+        // if voice commands feature is turned on then check if the user has microphone permission
+        if(ApplicationSettings.isVoiceCommandsEnabled(getContext())) {
+            checkAndRequestMicrophonePerms();
         }
     }
 
     @Override
     public void onViewCreated(@NonNull @NotNull View view, @Nullable @org.jetbrains.annotations.Nullable Bundle savedInstanceState) {
-        viewModel.getAimyboxState().observe(getViewLifecycleOwner(), state -> {
-            if (state == Aimybox.State.LISTENING) {
+        if(viewModel != null) {
+            viewModel.getAimyboxState().observe(getViewLifecycleOwner(), state -> {
+                if (state == Aimybox.State.LISTENING) {
+                    showVoiceTriggeredSnackbar(
+                            requireView(),
+                            requireActivity(),
+                            getContext(),
+                            requireActivity().findViewById(R.id.navigation),
+                            userHandler.getUser()
+                    );
 
-                showVoiceTriggeredSnackbar(
-                        requireView(),
-                        requireActivity(),
-                        getContext(),
-                        requireActivity().findViewById(R.id.navigation),
-                        userHandler.getUser()
-                );
+                }
 
-            }
-
-        });
+            });
+        }
     }
 
     @Override
@@ -359,6 +353,60 @@ public class DashboardFragment extends BaseInjectorFragment implements Coroutine
                             BaseTransientBottomBar.LENGTH_LONG).show();
                 }
             }
+        }
+        EasyPermissions.onRequestPermissionsResult(requestCode,permissions,grantResults,this);
+    }
+
+    private void initAimyboxViewModel(Context context){
+        if (viewModel == null) {
+            viewModel = new BaseAimyboxAssistantViewModel().getAimyboxAssistantViewModel(requireActivity(),
+                    new BaseAimybox().findAimyboxProvider(requireActivity()));
+            BaseAimybox.Companion.setInitialPhrase(context, getArguments(), viewModel);
+        }
+    }
+
+    public void checkAndRequestMicrophonePerms() {
+        String[] perms = {Manifest.permission.RECORD_AUDIO};
+
+        // if the user does not have permission, request it
+        if (!EasyPermissions.hasPermissions(requireContext(), perms)){
+
+            // Dialog requesting the user for microphone permission.
+            LOG.info("Microphone permissions not given, requesting");
+            EasyPermissions.requestPermissions(
+                    new PermissionRequest.Builder(this, RECORD_AUDIO_PERMISSION_REQ_CODE, perms)
+                            .setRationale(R.string.microphone_permission_voice_command)
+                            .setPositiveButtonText(R.string.grant_permission)
+                            .setNegativeButtonText(R.string.cancel)
+                            .setTheme(R.style.MaterialDialog)
+                            .build());
+        } else {
+            initAimyboxViewModel(requireContext());
+        }
+    }
+
+    @Override
+    public void onPermissionsGranted(int requestCode, @NonNull @NotNull List<String> perms) {
+        // if microphone permission is granted, initialise aimybox.
+        if (requestCode == RECORD_AUDIO_PERMISSION_REQ_CODE) {
+
+//            sendRecordingData();
+            // TODO init the aimybox? or ask for restart to get started with voice commands
+            showSnackbarLong(requireView(), getString(R.string.microphone_permission_granted));
+
+            initAimyboxViewModel(requireContext());
+        }
+    }
+
+    @Override
+    public void onPermissionsDenied(int requestCode, @NonNull @NotNull List<String> perms) {
+        // if permissions are not granted, show toast.
+        if (requestCode == RECORD_AUDIO_PERMISSION_REQ_CODE) {
+            // Disable the voice command feature .
+            ApplicationSettings.setVoiceCommandPreference(requireContext(), false);
+
+            // action opens app's general settings where user can grant microphone/any permission
+            showGrantMicrophonePermission(requireView(), requireContext(), requireActivity());
         }
     }
 
@@ -522,8 +570,8 @@ public class DashboardFragment extends BaseInjectorFragment implements Coroutine
             String[] perms;
             if (android.os.Build.VERSION.SDK_INT >= 31) {
                 perms = new String[]{
-                        Manifest.permission.ACCESS_FINE_LOCATION,
-                        Manifest.permission.ACCESS_COARSE_LOCATION
+                    Manifest.permission.ACCESS_FINE_LOCATION,
+                    Manifest.permission.ACCESS_COARSE_LOCATION
                 };
             }
             else {
@@ -534,9 +582,9 @@ public class DashboardFragment extends BaseInjectorFragment implements Coroutine
             ActivityCompat.requestPermissions(getActivity(), perms,
                     LOCATION_PERMISSION_REQUEST_CODE);
         } else if (user == null && ApplicationSettings.isTrackchunkUploadEnabled(getContext())) {
-            // cannot start, we need a user
-            LOG.info("cannot start, login is required for chunk upload feature");
-            Snackbar.make(getView(),
+                // cannot start, we need a user
+                LOG.info("cannot start, login is required for chunk upload feature");
+                Snackbar.make(getView(),
                     getString(R.string.dashboard_track_chunks_enabled_login),
                     Snackbar.LENGTH_LONG).show();
         } else if (user != null) {
@@ -544,29 +592,29 @@ public class DashboardFragment extends BaseInjectorFragment implements Coroutine
             if (ApplicationSettings.isTrackchunkUploadEnabled(getContext())) {
                 LOG.info("chunk upload is enabled, checking TermsOfUse");
                 mAgreementManager.verifyTermsOfUse(null, true)
-                        .subscribeOn(Schedulers.io())
-                        .observeOn(AndroidSchedulers.mainThread())
-                        .subscribe(termsOfUse -> {
-                            if (termsOfUse != null) {
-                                startRecording();
-                            } else {
-                                LOG.warn("No TermsOfUse received from verification");
+                    .subscribeOn(Schedulers.io())
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .subscribe(termsOfUse -> {
+                        if (termsOfUse != null) {
+                            startRecording();
+                        } else {
+                            LOG.warn("No TermsOfUse received from verification");
+                        }
+                    }, e -> {
+                        LOG.warn("Error during TermsOfUse verification", e);
+                        // inform the user about ToU acceptance
+                        Snackbar sb = Snackbar.make(getView(), String.format(getString(R.string.dashboard_accept_tou), getString(R.string.title_others)), Snackbar.LENGTH_INDEFINITE).setAction("OK", new View.OnClickListener() {
+                            @Override
+                            public void onClick(View v) {
+                                LOG.info("ToU Snackbar closed");
+                                Intent intent = new Intent(getActivity(), TermsOfUseActivity.class);
+                                getActivity().startActivity(intent);
                             }
-                        }, e -> {
-                            LOG.warn("Error during TermsOfUse verification", e);
-                            // inform the user about ToU acceptance
-                            Snackbar sb = Snackbar.make(getView(), String.format(getString(R.string.dashboard_accept_tou), getString(R.string.title_others)), Snackbar.LENGTH_INDEFINITE).setAction("OK", new View.OnClickListener() {
-                                @Override
-                                public void onClick(View v) {
-                                    LOG.info("ToU Snackbar closed");
-                                    Intent intent = new Intent(getActivity(), TermsOfUseActivity.class);
-                                    getActivity().startActivity(intent);
-                                }
-                            });
-                            Snackbar.SnackbarLayout layout = (Snackbar.SnackbarLayout) sb.getView();
-                            layout.setMinimumHeight(100);
-                            sb.show();
                         });
+                        Snackbar.SnackbarLayout layout = (Snackbar.SnackbarLayout) sb.getView();
+                        layout.setMinimumHeight(100);
+                        sb.show();
+                    });
             } else {
                 // we can check the ToUs later before upload
                 LOG.info("A user is logged in, chunk upload is disabled");
@@ -592,9 +640,9 @@ public class DashboardFragment extends BaseInjectorFragment implements Coroutine
                     Intent obdRecordingIntent = new Intent(getActivity(), RecordingService.class);
 
                     this.connectingDialog = DialogUtils.createProgressBarDialogBuilder(getContext(),
-                                    R.string.dashboard_connecting,
-                                    R.drawable.ic_bluetooth_white_24dp,
-                                    String.format(getString(R.string.dashboard_connecting_find_template), device.getName()))
+                            R.string.dashboard_connecting,
+                            R.drawable.ic_bluetooth_white_24dp,
+                            String.format(getString(R.string.dashboard_connecting_find_template), device.getName()))
                             .setNegativeButton(R.string.cancel, (dialog, which) -> {
                                 ServiceUtils.stopService(getActivity(), obdRecordingIntent);
                             })
